requirements/common.txt
Django==4.2.3
Pillow==10.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
python-dotenv
psycopg2-binary==2.9.9
six
django-storages[google]==1.14.2
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
django-cors-headers==4.3.1
django_filter==23.5
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
channels-postgres==1.0.4
django-modeltranslation==0.19.6
django-cities-light==3.10.1
